renamed exceptions package to failures as it contains strategies
