+++ cipher-suite/src/main/kotlin/net/corda/v5/cipher/suite/KeyEncodingService.kt
@@ -11,7 +11,7 @@ interface KeyEncodingService {
      * Decodes public key from byte array.
      *
      * @throws IllegalArgumentException if the key scheme is not supported.
-     * @throws net.corda.v5.crypto.exceptions.CryptoException for general cryptographic exceptions.
+     * @throws net.corda.v5.crypto.failures.CryptoException for general cryptographic exceptions.
      */
     fun decodePublicKey(encodedKey: ByteArray): PublicKey
 
@@ -19,7 +19,7 @@ interface KeyEncodingService {
      * Decodes public key from PEM encoded string.
      *
      * @throws IllegalArgumentException if the key scheme is not supported.
-     * @throws net.corda.v5.crypto.exceptions.CryptoException for general cryptographic exceptions.
+     * @throws net.corda.v5.crypto.failures.CryptoException for general cryptographic exceptions.
      */
     fun decodePublicKey(encodedKey: String): PublicKey
 
@@ -28,7 +28,7 @@ interface KeyEncodingService {
      * The default implementation returns the [PublicKey.getEncoded]
      *
      * @throws IllegalArgumentException if the key scheme is not supported.
-     * @throws net.corda.v5.crypto.exceptions.CryptoException for general cryptographic exceptions.
+     * @throws net.corda.v5.crypto.failures.CryptoException for general cryptographic exceptions.
      */
     fun encodeAsByteArray(publicKey: PublicKey): ByteArray = publicKey.encoded
 
@@ -36,7 +36,7 @@ interface KeyEncodingService {
      * Encodes public key to PEM encoded string.
      *
      * @throws IllegalArgumentException if the key scheme is not supported.
-     * @throws net.corda.v5.crypto.exceptions.CryptoException for general cryptographic exceptions.
+     * @throws net.corda.v5.crypto.failures.CryptoException for general cryptographic exceptions.
      */
     fun encodeAsString(publicKey: PublicKey): String
 
@@ -44,7 +44,7 @@ interface KeyEncodingService {
      * Convert a public key to a supported implementation. This can be used to convert a SUN's EC key to an BC key.
      *
      * @throws IllegalArgumentException if the key scheme is not supported.
-     * @throws net.corda.v5.crypto.exceptions.CryptoException for general cryptographic exceptions.
+     * @throws net.corda.v5.crypto.failures.CryptoException for general cryptographic exceptions.
      */
     fun toSupportedPublicKey(key: PublicKey): PublicKey
 }
