+++ crypto/src/main/kotlin/net/corda/v5/crypto/failures/CryptoExponentialThrottlingException.kt
@@ -1,4 +1,4 @@
-package net.corda.v5.crypto.exceptions
+package net.corda.v5.crypto.failures
 
 import net.corda.v5.base.annotations.CordaSerializable
 
@@ -8,26 +8,14 @@ import net.corda.v5.base.annotations.CordaSerializable
  */
 @CordaSerializable
 open class CryptoExponentialThrottlingException : CryptoThrottlingException {
-    companion object {
-        const val DEFAULT_THROTTLE_INITIAL_BACKOFF: Long = 1_000L
-        const val DEFAULT_THROTTLE_BACKOFF_MULTIPLIER: Long = 2
-        const val DEFAULT_THROTTLE_MAX_ATTEMPTS: Int = 5
-    }
-
-    private val initialBackoff: Long
-    private val backoffMultiplier: Long
-    private val maxAttempts: Int
+    private val backoffStrategy: CryptoExponentialRetryStrategy
 
     constructor(message: String) : super(message) {
-        this.initialBackoff = DEFAULT_THROTTLE_INITIAL_BACKOFF
-        this.backoffMultiplier = DEFAULT_THROTTLE_BACKOFF_MULTIPLIER
-        this.maxAttempts = DEFAULT_THROTTLE_MAX_ATTEMPTS
+        backoffStrategy = CryptoExponentialRetryStrategy()
     }
 
     constructor(message: String, cause: Throwable?) : super(message, cause) {
-        this.initialBackoff = DEFAULT_THROTTLE_INITIAL_BACKOFF
-        this.backoffMultiplier = DEFAULT_THROTTLE_BACKOFF_MULTIPLIER
-        this.maxAttempts = DEFAULT_THROTTLE_MAX_ATTEMPTS
+        backoffStrategy = CryptoExponentialRetryStrategy()
     }
 
     constructor(
@@ -37,9 +25,11 @@ open class CryptoExponentialThrottlingException : CryptoThrottlingException {
         maxAttempts: Int,
         cause: Throwable?
     ) : super(message, cause) {
-        this.initialBackoff = initialBackoff
-        this.backoffMultiplier = backoffMultiplier
-        this.maxAttempts = maxAttempts
+        backoffStrategy = CryptoExponentialRetryStrategy(
+            maxAttempts = maxAttempts,
+            initialBackoff = initialBackoff,
+            backoffMultiplier = backoffMultiplier
+        )
     }
 
     /**
@@ -52,11 +42,5 @@ open class CryptoExponentialThrottlingException : CryptoThrottlingException {
      * and then give up.
      */
     override fun getBackoff(attempt: Int, currentBackoffMillis: Long): Long =
-        if(attempt < 1 || attempt > maxAttempts) {
-            -1
-        } else if(attempt == 1) {
-            initialBackoff
-        } else {
-            currentBackoffMillis * backoffMultiplier
-        }
+        backoffStrategy.getBackoff(attempt, currentBackoffMillis)
 }
