+++ crypto/src/main/kotlin/net/corda/v5/crypto/failures/CryptoRetryException.kt
@@ -1,12 +1,12 @@
-package net.corda.v5.crypto.exceptions
+package net.corda.v5.crypto.failures
 
 import net.corda.v5.base.annotations.CordaSerializable
 
 /**
- * Signals that the signature verification has failed.
+ * Signals that the transient fault handling was attempted but wasn't successful.
  */
 @CordaSerializable
-class CryptoSignatureException : CryptoException {
+class CryptoRetryException : CryptoException {
     constructor(message: String) : super(message)
 
     constructor(message: String, cause: Throwable?) : super(message, cause)
