+++ crypto/src/main/kotlin/net/corda/v5/crypto/failures/CryptoSignatureException.kt
@@ -1,12 +1,12 @@
-package net.corda.v5.crypto.exceptions
+package net.corda.v5.crypto.failures
 
 import net.corda.v5.base.annotations.CordaSerializable
 
 /**
- * Signals that the transient fault handling was attempted but wasn't successful.
+ * Signals that the signature verification has failed.
  */
 @CordaSerializable
-class CryptoRetryException : CryptoException {
+class CryptoSignatureException : CryptoException {
     constructor(message: String) : super(message)
 
     constructor(message: String, cause: Throwable?) : super(message, cause)
